LPD-66925 Move common logic to base class
