+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/internal/layout/display/page/BlogsLayoutDisplayPageProvider.java
@@ -16,19 +16,12 @@ import com.liferay.info.item.InfoItemReference;
 import com.liferay.layout.display.page.BaseLayoutDisplayPageProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageObjectProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageProvider;
-import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.portlet.constants.FriendlyURLResolverConstants;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.util.GroupThreadLocal;
-import com.liferay.portal.kernel.util.Validator;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -62,25 +55,6 @@ public class BlogsLayoutDisplayPageProvider
 
 		return new BlogsLayoutDisplayPageObjectProvider(
 			_assetHelper, blogsEntry, _infoItemFriendlyURLProvider, _language);
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<BlogsEntry>
-		getLayoutDisplayPageObjectProvider(
-			InfoItemReference infoItemReference) {
-
-		return getLayoutDisplayPageObjectProvider(
-			_getGroupId(), infoItemReference);
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<BlogsEntry>
-		getLayoutDisplayPageObjectProvider(
-			long groupId, InfoItemReference infoItemReference) {
-
-		return _getLayoutDisplayPageObjectProvider(
-			_getGroupId(groupId, infoItemReference.getInfoItemIdentifier()),
-			infoItemReference);
 	}
 
 	@Override
@@ -119,74 +93,9 @@ public class BlogsLayoutDisplayPageProvider
 			_assetHelper, blogsEntry, _infoItemFriendlyURLProvider, _language);
 	}
 
-	private long _getCompanyId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getCompanyId();
-		}
-
-		Long companyId = CompanyThreadLocal.getCompanyId();
-
-		if (companyId != null) {
-			return companyId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor company thread local " +
-				"are initialized");
-	}
-
-	private long _getGroupId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getScopeGroupId();
-		}
-
-		Long groupId = GroupThreadLocal.getGroupId();
-
-		if (groupId != null) {
-			return groupId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor group thread local are " +
-				"initialized");
-	}
-
-	private long _getGroupId(
-		long groupId, InfoItemIdentifier infoItemIdentifier) {
-
-		try {
-			if (!(infoItemIdentifier instanceof ERCInfoItemIdentifier)) {
-				return groupId;
-			}
-
-			ERCInfoItemIdentifier ercInfoItemIdentifier =
-				(ERCInfoItemIdentifier)infoItemIdentifier;
-
-			if (Validator.isNull(
-					ercInfoItemIdentifier.getScopeExternalReferenceCode())) {
-
-				return groupId;
-			}
-
-			Group group = GroupLocalServiceUtil.getGroupByExternalReferenceCode(
-				ercInfoItemIdentifier.getScopeExternalReferenceCode(),
-				_getCompanyId());
-
-			return group.getGroupId();
-		}
-		catch (PortalException portalException) {
-			return ReflectionUtil.throwException(portalException);
-		}
-	}
-
-	private BlogsLayoutDisplayPageObjectProvider
-		_getLayoutDisplayPageObjectProvider(
+	@Override
+	protected BlogsLayoutDisplayPageObjectProvider
+		doGetLayoutDisplayPageObjectProvider(
 			long groupId, InfoItemReference infoItemReference) {
 
 		InfoItemIdentifier infoItemIdentifier =
