+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/layout/display/page/FileEntryLayoutDisplayPageProvider.java
@@ -16,7 +16,6 @@ import com.liferay.info.item.InfoItemReference;
 import com.liferay.layout.display.page.BaseLayoutDisplayPageProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageObjectProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageProvider;
-import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
@@ -27,11 +26,7 @@ import com.liferay.portal.kernel.repository.RepositoryProvider;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.GroupThreadLocal;
 import com.liferay.portal.kernel.util.Validator;
 
 import org.osgi.service.component.annotations.Component;
@@ -64,25 +59,6 @@ public class FileEntryLayoutDisplayPageProvider
 
 		return new FileEntryLayoutDisplayPageObjectProvider(
 			fileEntry, _infoItemFriendlyURLProvider, _language);
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<FileEntry>
-		getLayoutDisplayPageObjectProvider(
-			InfoItemReference infoItemReference) {
-
-		return getLayoutDisplayPageObjectProvider(
-			_getGroupId(), infoItemReference);
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<FileEntry>
-		getLayoutDisplayPageObjectProvider(
-			long groupId, InfoItemReference infoItemReference) {
-
-		return _getLayoutDisplayPageObjectProvider(
-			_getGroupId(groupId, infoItemReference.getInfoItemIdentifier()),
-			infoItemReference);
 	}
 
 	@Override
@@ -123,74 +99,9 @@ public class FileEntryLayoutDisplayPageProvider
 				FileEntry.class.getName(), GetterUtil.getLong(urlTitle)));
 	}
 
-	private long _getCompanyId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getCompanyId();
-		}
-
-		Long companyId = CompanyThreadLocal.getCompanyId();
-
-		if (companyId != null) {
-			return companyId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor company thread local " +
-				"are initialized");
-	}
-
-	private long _getGroupId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getScopeGroupId();
-		}
-
-		Long groupId = GroupThreadLocal.getGroupId();
-
-		if (groupId != null) {
-			return groupId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor group thread local are " +
-				"initialized");
-	}
-
-	private long _getGroupId(
-		long groupId, InfoItemIdentifier infoItemIdentifier) {
-
-		try {
-			if (!(infoItemIdentifier instanceof ERCInfoItemIdentifier)) {
-				return groupId;
-			}
-
-			ERCInfoItemIdentifier ercInfoItemIdentifier =
-				(ERCInfoItemIdentifier)infoItemIdentifier;
-
-			if (Validator.isNull(
-					ercInfoItemIdentifier.getScopeExternalReferenceCode())) {
-
-				return groupId;
-			}
-
-			Group group = GroupLocalServiceUtil.getGroupByExternalReferenceCode(
-				ercInfoItemIdentifier.getScopeExternalReferenceCode(),
-				_getCompanyId());
-
-			return group.getGroupId();
-		}
-		catch (PortalException portalException) {
-			return ReflectionUtil.throwException(portalException);
-		}
-	}
-
-	private FileEntryLayoutDisplayPageObjectProvider
-		_getLayoutDisplayPageObjectProvider(
+	@Override
+	protected FileEntryLayoutDisplayPageObjectProvider
+		doGetLayoutDisplayPageObjectProvider(
 			long groupId, InfoItemReference infoItemReference) {
 
 		try {
