+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/layout/display/page/JournalArticleLayoutDisplayPageProvider.java
@@ -16,18 +16,11 @@ import com.liferay.journal.service.JournalArticleLocalService;
 import com.liferay.layout.display.page.BaseLayoutDisplayPageProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageObjectProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageProvider;
-import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.portlet.constants.FriendlyURLResolverConstants;
-import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.GroupThreadLocal;
 import com.liferay.portal.kernel.util.Validator;
 
 import org.osgi.service.component.annotations.Component;
@@ -55,15 +48,6 @@ public class JournalArticleLayoutDisplayPageProvider
 
 	@Override
 	public LayoutDisplayPageObjectProvider<JournalArticle>
-		getLayoutDisplayPageObjectProvider(
-			InfoItemReference infoItemReference) {
-
-		return _getLayoutDisplayPageObjectProvider(
-			_getGroupId(), infoItemReference);
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<JournalArticle>
 		getLayoutDisplayPageObjectProvider(JournalArticle article) {
 
 		try {
@@ -77,16 +61,6 @@ public class JournalArticleLayoutDisplayPageProvider
 		catch (PortalException portalException) {
 			throw new RuntimeException(portalException);
 		}
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<JournalArticle>
-		getLayoutDisplayPageObjectProvider(
-			long groupId, InfoItemReference infoItemReference) {
-
-		return _getLayoutDisplayPageObjectProvider(
-			_getGroupId(groupId, infoItemReference.getInfoItemIdentifier()),
-			infoItemReference);
 	}
 
 	@Override
@@ -128,110 +102,9 @@ public class JournalArticleLayoutDisplayPageProvider
 		}
 	}
 
-	@Reference
-	protected AssetHelper assetHelper;
-
-	@Reference
-	protected JournalArticleLocalService journalArticleLocalService;
-
-	@Reference
-	protected SiteConnectedGroupGroupProvider siteConnectedGroupGroupProvider;
-
-	private JournalArticle _getArticle(
-			long groupId, String urlTitle, String version)
-		throws PortalException {
-
-		for (long connectedGroupId :
-				siteConnectedGroupGroupProvider.
-					getCurrentAndAncestorSiteAndDepotGroupIds(groupId, true)) {
-
-			JournalArticle article = null;
-
-			if (Validator.isNotNull(version)) {
-				article = journalArticleLocalService.fetchArticleByUrlTitle(
-					connectedGroupId, urlTitle, GetterUtil.getDouble(version));
-			}
-			else {
-				article = journalArticleLocalService.fetchArticleByUrlTitle(
-					connectedGroupId, urlTitle);
-			}
-
-			if (article != null) {
-				return article;
-			}
-		}
-
-		return null;
-	}
-
-	private long _getCompanyId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getCompanyId();
-		}
-
-		Long companyId = CompanyThreadLocal.getCompanyId();
-
-		if (companyId != null) {
-			return companyId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor company thread local " +
-				"are initialized");
-	}
-
-	private long _getGroupId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getScopeGroupId();
-		}
-
-		Long groupId = GroupThreadLocal.getGroupId();
-
-		if (groupId != null) {
-			return groupId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor group thread local are " +
-				"initialized");
-	}
-
-	private long _getGroupId(
-		long groupId, InfoItemIdentifier infoItemIdentifier) {
-
-		try {
-			if (!(infoItemIdentifier instanceof ERCInfoItemIdentifier)) {
-				return groupId;
-			}
-
-			ERCInfoItemIdentifier ercInfoItemIdentifier =
-				(ERCInfoItemIdentifier)infoItemIdentifier;
-
-			if (Validator.isNull(
-					ercInfoItemIdentifier.getScopeExternalReferenceCode())) {
-
-				return groupId;
-			}
-
-			Group group = GroupLocalServiceUtil.getGroupByExternalReferenceCode(
-				ercInfoItemIdentifier.getScopeExternalReferenceCode(),
-				_getCompanyId());
-
-			return group.getGroupId();
-		}
-		catch (PortalException portalException) {
-			return ReflectionUtil.throwException(portalException);
-		}
-	}
-
-	private JournalArticleLayoutDisplayPageObjectProvider
-		_getLayoutDisplayPageObjectProvider(
+	@Override
+	protected JournalArticleLayoutDisplayPageObjectProvider
+		doGetLayoutDisplayPageObjectProvider(
 			long groupId, InfoItemReference infoItemReference) {
 
 		InfoItemIdentifier infoItemIdentifier =
@@ -281,6 +154,42 @@ public class JournalArticleLayoutDisplayPageProvider
 		catch (PortalException portalException) {
 			throw new RuntimeException(portalException);
 		}
+	}
+
+	@Reference
+	protected AssetHelper assetHelper;
+
+	@Reference
+	protected JournalArticleLocalService journalArticleLocalService;
+
+	@Reference
+	protected SiteConnectedGroupGroupProvider siteConnectedGroupGroupProvider;
+
+	private JournalArticle _getArticle(
+			long groupId, String urlTitle, String version)
+		throws PortalException {
+
+		for (long connectedGroupId :
+				siteConnectedGroupGroupProvider.
+					getCurrentAndAncestorSiteAndDepotGroupIds(groupId, true)) {
+
+			JournalArticle article = null;
+
+			if (Validator.isNotNull(version)) {
+				article = journalArticleLocalService.fetchArticleByUrlTitle(
+					connectedGroupId, urlTitle, GetterUtil.getDouble(version));
+			}
+			else {
+				article = journalArticleLocalService.fetchArticleByUrlTitle(
+					connectedGroupId, urlTitle);
+			}
+
+			if (article != null) {
+				return article;
+			}
+		}
+
+		return null;
 	}
 
 	private boolean _isShow(JournalArticle article) {
