+++ modules/apps/knowledge-base/knowledge-base-web/src/main/java/com/liferay/knowledge/base/web/internal/layout/display/page/KBArticleLayoutDisplayPageProvider.java
@@ -16,18 +16,10 @@ import com.liferay.knowledge.base.service.KBArticleLocalService;
 import com.liferay.layout.display.page.BaseLayoutDisplayPageProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageObjectProvider;
 import com.liferay.layout.display.page.LayoutDisplayPageProvider;
-import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringUtil;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.portlet.constants.FriendlyURLResolverConstants;
-import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
-import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.util.GroupThreadLocal;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
 import java.util.List;
@@ -55,15 +47,6 @@ public class KBArticleLayoutDisplayPageProvider
 
 	@Override
 	public LayoutDisplayPageObjectProvider<KBArticle>
-		getLayoutDisplayPageObjectProvider(
-			InfoItemReference infoItemReference) {
-
-		return getLayoutDisplayPageObjectProvider(
-			_getGroupId(), infoItemReference);
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<KBArticle>
 		getLayoutDisplayPageObjectProvider(KBArticle kbArticle) {
 
 		try {
@@ -81,16 +64,6 @@ public class KBArticleLayoutDisplayPageProvider
 		catch (PortalException portalException) {
 			throw new RuntimeException(portalException);
 		}
-	}
-
-	@Override
-	public LayoutDisplayPageObjectProvider<KBArticle>
-		getLayoutDisplayPageObjectProvider(
-			long groupId, InfoItemReference infoItemReference) {
-
-		return _getLayoutDisplayPageObjectProvider(
-			_getGroupId(groupId, infoItemReference.getInfoItemIdentifier()),
-			infoItemReference);
 	}
 
 	@Override
@@ -129,89 +102,9 @@ public class KBArticleLayoutDisplayPageProvider
 		}
 	}
 
-	private long _getCompanyId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getCompanyId();
-		}
-
-		Long companyId = CompanyThreadLocal.getCompanyId();
-
-		if (companyId != null) {
-			return companyId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor company thread local " +
-				"are initialized");
-	}
-
-	private long _getGroupId() {
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if (serviceContext != null) {
-			return serviceContext.getScopeGroupId();
-		}
-
-		Long groupId = GroupThreadLocal.getGroupId();
-
-		if (groupId != null) {
-			return groupId;
-		}
-
-		throw new IllegalStateException(
-			"Neither service context thread local nor group thread local are " +
-				"initialized");
-	}
-
-	private long _getGroupId(
-		long groupId, InfoItemIdentifier infoItemIdentifier) {
-
-		try {
-			if (!(infoItemIdentifier instanceof ERCInfoItemIdentifier)) {
-				return groupId;
-			}
-
-			ERCInfoItemIdentifier ercInfoItemIdentifier =
-				(ERCInfoItemIdentifier)infoItemIdentifier;
-
-			if (Validator.isNull(
-					ercInfoItemIdentifier.getScopeExternalReferenceCode())) {
-
-				return groupId;
-			}
-
-			Group group = GroupLocalServiceUtil.getGroupByExternalReferenceCode(
-				ercInfoItemIdentifier.getScopeExternalReferenceCode(),
-				_getCompanyId());
-
-			return group.getGroupId();
-		}
-		catch (PortalException portalException) {
-			return ReflectionUtil.throwException(portalException);
-		}
-	}
-
-	private long _getKBFolderId(long groupId, List<String> urlTitleParts) {
-		if (urlTitleParts.size() == 1) {
-			return KBFolderConstants.DEFAULT_PARENT_FOLDER_ID;
-		}
-
-		KBArticle kbArticle = _kbArticleLocalService.fetchKBArticleByUrlTitle(
-			groupId, urlTitleParts.get(0), urlTitleParts.get(1));
-
-		if (kbArticle != null) {
-			return kbArticle.getKbFolderId();
-		}
-
-		return KBFolderConstants.DEFAULT_PARENT_FOLDER_ID;
-	}
-
-	private KBArticleLayoutDisplayPageObjectProvider
-		_getLayoutDisplayPageObjectProvider(
+	@Override
+	protected KBArticleLayoutDisplayPageObjectProvider
+		doGetLayoutDisplayPageObjectProvider(
 			long groupId, InfoItemReference infoItemReference) {
 
 		try {
@@ -261,6 +154,21 @@ public class KBArticleLayoutDisplayPageProvider
 		catch (PortalException portalException) {
 			throw new RuntimeException(portalException);
 		}
+	}
+
+	private long _getKBFolderId(long groupId, List<String> urlTitleParts) {
+		if (urlTitleParts.size() == 1) {
+			return KBFolderConstants.DEFAULT_PARENT_FOLDER_ID;
+		}
+
+		KBArticle kbArticle = _kbArticleLocalService.fetchKBArticleByUrlTitle(
+			groupId, urlTitleParts.get(0), urlTitleParts.get(1));
+
+		if (kbArticle != null) {
+			return kbArticle.getKbFolderId();
+		}
+
+		return KBFolderConstants.DEFAULT_PARENT_FOLDER_ID;
 	}
 
 	@Reference
